remove old outbound line edit modal
