+++ client/packages/invoices/src/OutboundShipment/DetailView/OutboundLineEdit/hooks/useOpenedWithBarcode.ts
