--- client/packages/invoices/src/OutboundShipment/DetailView/OutboundLineEditModal/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './useDraftOutboundLines';
-export * from './useNextItem';
-export * from './useOutboundLineEditRows';
