--- client/packages/invoices/src/OutboundShipment/DetailView/OutboundLineEditModal/index.ts
@@ -1,2 +1,0 @@
-export * from './OutboundLineEdit';
-export * from './hooks';
